app/tools: extract catalog definition helpers
