+++ crates/app/src/tools/catalog.rs
@@ -15,6 +15,50 @@ use crate::conversation::ConstrainedSubagentContractView;
 #[cfg(test)]
 use crate::conversation::ConstrainedSubagentProfile;
 
+#[path = "catalog_metadata_support.rs"]
+mod metadata_support;
+use metadata_support::{
+    tool_argument_hint, tool_parameter_types, tool_required_fields, tool_search_hint, tool_tags,
+};
+#[path = "catalog_core_definition_support.rs"]
+mod core_definition_support;
+use core_definition_support::{
+    direct_browser_definition, direct_exec_definition, direct_memory_definition,
+    direct_read_definition, direct_web_definition, direct_write_definition, tool_invoke_definition,
+    tool_search_definition,
+};
+#[path = "catalog_browser_definition_support.rs"]
+mod browser_definition_support;
+use browser_definition_support::{
+    browser_click_definition, browser_companion_click_definition,
+    browser_companion_navigate_definition, browser_companion_session_start_definition,
+    browser_companion_session_stop_definition, browser_companion_snapshot_definition,
+    browser_companion_type_definition, browser_companion_wait_definition,
+    browser_extract_definition, browser_open_definition,
+};
+#[path = "catalog_external_skills_definition_support.rs"]
+mod external_skills_definition_support;
+use external_skills_definition_support::{
+    config_import_definition, external_skills_fetch_definition, external_skills_inspect_definition,
+    external_skills_install_definition, external_skills_invoke_definition,
+    external_skills_list_definition, external_skills_policy_definition,
+    external_skills_recommend_definition, external_skills_remove_definition,
+    external_skills_resolve_definition, external_skills_search_definition,
+    external_skills_source_search_definition, provider_switch_definition,
+};
+#[path = "catalog_session_definition_support.rs"]
+mod session_definition_support;
+use session_definition_support::{
+    approval_request_resolve_definition, approval_request_status_definition,
+    approval_requests_list_definition, delegate_async_definition, delegate_definition,
+    session_archive_definition, session_cancel_definition, session_continue_definition,
+    session_events_definition, session_recover_definition, session_search_definition,
+    session_status_definition, session_tool_policy_clear_definition,
+    session_tool_policy_set_definition, session_tool_policy_status_definition,
+    session_wait_definition, sessions_history_definition, sessions_list_definition,
+    sessions_send_definition,
+};
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
 pub enum ToolExecutionKind {
     Core,
@@ -2280,1154 +2324,6 @@ fn tool_visibility_gate_enabled_for_runtime_policy(
     }
 }
 
-fn tool_search_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Discover a specialized tool when the direct tools do not fit the task.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "type": "string",
-                        "description": "Short capability phrase for the specialized tool you need. Match the prompt snippets shown in the system prompt."
-                    },
-                    "exact_tool_id": {
-                        "type": "string",
-                        "description": "Optional exact tool id to refresh a known visible tool card."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "description": "Optional maximum number of search results to return."
-                    }
-                },
-                "required": [],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn direct_read_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "path": {
-                        "type": "string",
-                        "description": "Read one file at this workspace-relative or absolute path."
-                    },
-                    "max_bytes": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 8_388_608,
-                        "description": "Optional read limit in bytes when reading one file or file window."
-                    },
-                    "offset": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional 1-indexed line number to start from when reading one file."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional maximum number of lines to return when reading one file."
-                    },
-                    "query": {
-                        "type": "string",
-                        "description": "Search workspace file contents for this text."
-                    },
-                    "pattern": {
-                        "type": "string",
-                        "description": "List workspace paths that match this glob pattern."
-                    },
-                    "root": {
-                        "type": "string",
-                        "description": "Optional search root path for query or pattern mode."
-                    },
-                    "glob": {
-                        "type": "string",
-                        "description": "Optional file glob filter applied only in query mode."
-                    },
-                    "max_results": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 200,
-                        "description": "Optional maximum result count for query or pattern mode."
-                    },
-                    "max_bytes_per_file": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 1_048_576,
-                        "description": "Optional per-file scan budget used only in query mode."
-                    },
-                    "case_sensitive": {
-                        "type": "boolean",
-                        "description": "Use case-sensitive matching in query mode. Defaults to false."
-                    },
-                    "include_directories": {
-                        "type": "boolean",
-                        "description": "Include matching directories in pattern mode. Defaults to false."
-                    }
-                },
-                "anyOf": [
-                    {
-                        "required": ["path"]
-                    },
-                    {
-                        "required": ["query"]
-                    },
-                    {
-                        "required": ["pattern"]
-                    }
-                ],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn exact_edit_block_definition() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "old_text": {
-                "type": "string",
-                "minLength": 1,
-                "description": "Exact text for one targeted replacement. It must match uniquely in the original file and must not overlap any other edit block."
-            },
-            "new_text": {
-                "type": "string",
-                "description": "Replacement text for this targeted edit block."
-            }
-        },
-        "required": ["old_text", "new_text"],
-        "additionalProperties": false
-    })
-}
-
-fn direct_write_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "path": {
-                        "type": "string",
-                        "description": "Target file path."
-                    },
-                    "content": {
-                        "type": "string",
-                        "description": "Whole-file content used for create or replace mode."
-                    },
-                    "create_dirs": {
-                        "type": "boolean",
-                        "description": "Create parent directories when missing. Defaults to true."
-                    },
-                    "overwrite": {
-                        "type": "boolean",
-                        "description": "Allow replacing an existing file. Defaults to false."
-                    },
-                    "edits": {
-                        "type": "array",
-                        "description": "One or more exact text replacement blocks matched against the original file. Merge nearby edits instead of sending overlapping blocks.",
-                        "items": exact_edit_block_definition(),
-                        "minItems": 1
-                    },
-                    "old_string": {
-                        "type": "string",
-                        "minLength": 1,
-                        "description": "Legacy single-block exact edit field. Prefer `edits` for new requests."
-                    },
-                    "new_string": {
-                        "type": "string",
-                        "description": "Legacy replacement text paired with `old_string`. Prefer `edits` for new requests."
-                    },
-                    "replace_all": {
-                        "type": "boolean",
-                        "description": "Legacy single-block mode only. Replace all matches instead of requiring a unique match. Defaults to false."
-                    }
-                },
-                "required": ["path"],
-                "anyOf": [
-                    {
-                        "required": ["content"]
-                    },
-                    {
-                        "required": ["edits"]
-                    },
-                    {
-                        "required": ["old_string", "new_string"]
-                    }
-                ],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn direct_exec_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "command": {
-                        "type": "string",
-                        "description": "Executable command or simple shell command. Routes to argv mode unless it clearly uses shell syntax."
-                    },
-                    "script": {
-                        "type": "string",
-                        "description": "Raw shell or bash script text. Use this for pipes, redirects, chaining, or multi-line commands."
-                    },
-                    "args": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                        "description": "Optional command arguments for argv mode."
-                    },
-                    "timeout_ms": {
-                        "type": "integer",
-                        "minimum": 1000,
-                        "maximum": 600000,
-                        "description": "Optional command timeout in milliseconds."
-                    },
-                    "cwd": {
-                        "type": "string",
-                        "description": "Optional working directory."
-                    }
-                },
-                "anyOf": [
-                    {
-                        "required": ["command"]
-                    },
-                    {
-                        "required": ["script"]
-                    }
-                ],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn direct_web_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "url": {
-                        "type": "string",
-                        "description": "Fetch or request this HTTP or HTTPS URL without using a web-search provider."
-                    },
-                    "mode": {
-                        "type": "string",
-                        "enum": ["readable_text", "raw_text"],
-                        "description": "Fetch rendering mode. Used only for plain fetch mode."
-                    },
-                    "max_bytes": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_WEB_FETCH_MAX_BYTES,
-                        "description": "Optional response byte limit."
-                    },
-                    "query": {
-                        "type": "string",
-                        "description": "Search the public web for this query through web-search providers. This is separate from plain URL fetch/request mode."
-                    },
-                    "provider": {
-                        "type": "string",
-                        "enum": crate::config::WEB_SEARCH_PROVIDER_SCHEMA_VALUES,
-                        "description": crate::config::web_search_provider_parameter_description()
-                    },
-                    "max_results": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 10,
-                        "description": "Optional maximum result count in search mode."
-                    },
-                    "method": {
-                        "type": "string",
-                        "description": "Optional HTTP method. When present, web routes to low-level request mode."
-                    },
-                    "headers": {
-                        "type": "object",
-                        "additionalProperties": {"type": "string"},
-                        "description": "Optional HTTP headers for request mode."
-                    },
-                    "body": {
-                        "type": "string",
-                        "description": "Optional request body for request mode."
-                    },
-                    "content_type": {
-                        "type": "string",
-                        "description": "Optional Content-Type header for request mode."
-                    }
-                },
-                "anyOf": [
-                    {
-                        "required": ["url"]
-                    },
-                    {
-                        "required": ["query"]
-                    }
-                ],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn direct_browser_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "url": {
-                        "type": "string",
-                        "description": "Open a page or continue a browser session at this HTTP or HTTPS URL."
-                    },
-                    "max_bytes": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_WEB_FETCH_MAX_BYTES,
-                        "description": "Optional byte limit used when opening a page."
-                    },
-                    "session_id": {
-                        "type": "string",
-                        "description": "Existing browser session identifier for follow-up reads or interactions."
-                    },
-                    "mode": {
-                        "type": "string",
-                        "enum": ["page_text", "title", "links", "selector_text", "summary", "html"],
-                        "description": "Read mode for browser inspection."
-                    },
-                    "selector": {
-                        "type": "string",
-                        "description": "CSS selector for focused extraction or browser interaction."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_BROWSER_MAX_LINKS,
-                        "description": "Maximum extracted items when the browser result returns a list."
-                    },
-                    "link_id": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_BROWSER_MAX_LINKS,
-                        "description": "One-based link identifier returned by the current page snapshot."
-                    },
-                    "text": {
-                        "type": "string",
-                        "description": "Text to type into the selected element."
-                    },
-                    "condition": {
-                        "type": "string",
-                        "description": "Optional wait condition for browser session progress."
-                    },
-                    "timeout_ms": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 30000,
-                        "description": "Optional wait timeout in milliseconds."
-                    }
-                },
-                "anyOf": [
-                    {
-                        "required": ["url"]
-                    },
-                    {
-                        "required": ["session_id"]
-                    },
-                    {
-                        "required": ["session_id", "link_id"]
-                    },
-                    {
-                        "required": ["session_id", "selector"]
-                    },
-                    {
-                        "required": ["session_id", "selector", "text"]
-                    },
-                    {
-                        "required": ["session_id", "url"]
-                    }
-                ],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn direct_memory_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "type": "string",
-                        "description": "Search durable memory and canonical recall for this query."
-                    },
-                    "max_results": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 8,
-                        "description": "Optional maximum number of memory hits to return."
-                    },
-                    "path": {
-                        "type": "string",
-                        "description": "Read one durable memory file at this path."
-                    },
-                    "from": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional 1-based starting line number for path mode."
-                    },
-                    "lines": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 200,
-                        "description": "Optional number of lines to read in path mode."
-                    }
-                },
-                "anyOf": [
-                    {
-                        "required": ["query"]
-                    },
-                    {
-                        "required": ["path"]
-                    }
-                ],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_open_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "url": {
-                        "type": "string",
-                        "description": "HTTP or HTTPS URL to open."
-                    },
-                    "max_bytes": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_WEB_FETCH_MAX_BYTES,
-                        "description": "Optional per-call read limit in bytes. Cannot exceed the configured runtime max."
-                    }
-                },
-                "required": ["url"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_extract_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Bounded browser session identifier returned by browser.open."
-                    },
-                    "mode": {
-                        "type": "string",
-                        "enum": ["page_text", "title", "links", "selector_text"],
-                        "description": "Extraction mode. Defaults to `page_text`."
-                    },
-                    "selector": {
-                        "type": "string",
-                        "description": "Optional CSS selector used only with `selector_text` mode."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_BROWSER_MAX_LINKS,
-                        "description": "Maximum extracted items when the mode returns a list."
-                    }
-                },
-                "required": ["session_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_click_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Bounded browser session identifier returned by browser.open."
-                    },
-                    "link_id": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": crate::config::MAX_BROWSER_MAX_LINKS,
-                        "description": "One-based link identifier returned in the current page snapshot."
-                    }
-                },
-                "required": ["session_id", "link_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_session_start_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "url": {
-                        "type": "string",
-                        "description": "HTTP or HTTPS URL to open in the managed browser companion session."
-                    }
-                },
-                "required": ["url"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_navigate_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Browser companion session identifier returned by browser.companion.session.start."
-                    },
-                    "url": {
-                        "type": "string",
-                        "description": "HTTP or HTTPS URL to load next."
-                    }
-                },
-                "required": ["session_id", "url"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_snapshot_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Browser companion session identifier returned by browser.companion.session.start."
-                    },
-                    "mode": {
-                        "type": "string",
-                        "enum": ["summary", "html", "links"],
-                        "description": "Optional snapshot mode. Defaults to `summary`."
-                    }
-                },
-                "required": ["session_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_wait_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Browser companion session identifier returned by browser.companion.session.start."
-                    },
-                    "condition": {
-                        "type": "string",
-                        "description": "Optional companion-side wait condition."
-                    },
-                    "timeout_ms": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 30000,
-                        "description": "Optional maximum wait in milliseconds."
-                    }
-                },
-                "required": ["session_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_session_stop_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Browser companion session identifier returned by browser.companion.session.start."
-                    }
-                },
-                "required": ["session_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_click_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Browser companion session identifier returned by browser.companion.session.start."
-                    },
-                    "selector": {
-                        "type": "string",
-                        "description": "Selector for the element to click."
-                    }
-                },
-                "required": ["session_id", "selector"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn browser_companion_type_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Browser companion session identifier returned by browser.companion.session.start."
-                    },
-                    "selector": {
-                        "type": "string",
-                        "description": "Selector for the element to type into."
-                    },
-                    "text": {
-                        "type": "string",
-                        "description": "Text to enter."
-                    }
-                },
-                "required": ["session_id", "selector", "text"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn tool_invoke_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Invoke a discovered non-core tool using a valid lease from tool_search.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "tool_id": {
-                        "type": "string",
-                        "description": "Canonical id of the discovered tool."
-                    },
-                    "lease": {
-                        "type": "string",
-                        "description": "Short-lived lease returned by tool_search."
-                    },
-                    "arguments": {
-                        "type": "object",
-                        "description": "Arguments for the discovered tool payload."
-                    }
-                },
-                "required": ["tool_id", "lease", "arguments"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn config_import_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Import, discover, plan, merge, apply, and roll back legacy agent workspace config and related external-skills state into native Loong config.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "input_path": {
-                        "type": "string",
-                        "description": "Path to the legacy agent workspace, config root, or portable import file. Required for all modes except rollback_last_apply."
-                    },
-                    "mode": {
-                        "type": "string",
-                        "enum": [
-                            "plan",
-                            "apply",
-                            "discover",
-                            "plan_many",
-                            "recommend_primary",
-                            "merge_profiles",
-                            "map_external_skills",
-                            "apply_selected",
-                            "rollback_last_apply"
-                        ],
-                        "description": "Migration mode. Defaults to `plan` when omitted."
-                    },
-                    "source": {
-                        "type": "string",
-                        "enum": ["auto", "nanobot", "openclaw", "picoclaw", "zeroclaw", "nanoclaw"],
-                        "description": "Optional claw-family source hint for plan/apply modes. Defaults to automatic detection."
-                    },
-                    "source_id": {
-                        "type": "string",
-                        "description": "Selected source identifier for apply_selected mode."
-                    },
-                    "selection_id": {
-                        "type": "string",
-                        "description": "Alias of source_id for apply_selected mode."
-                    },
-                    "primary_source_id": {
-                        "type": "string",
-                        "description": "Primary source identifier for safe profile merge in apply_selected mode."
-                    },
-                    "primary_selection_id": {
-                        "type": "string",
-                        "description": "Alias of primary_source_id for safe profile merge in apply_selected mode."
-                    },
-                    "safe_profile_merge": {
-                        "type": "boolean",
-                        "description": "Enable safe multi-source profile merge in apply_selected mode."
-                    },
-                    "apply_external_skills_plan": {
-                        "type": "boolean",
-                        "description": "When true, apply a generated external-skills mapping addendum into profile_note during apply_selected."
-                    },
-                    "output_path": {
-                        "type": "string",
-                        "description": "Optional target config path. In plan, when present, config.import reads this path to preview the merged result. Required in apply/apply_selected/rollback_last_apply modes."
-                    },
-                    "force": {
-                        "type": "boolean",
-                        "description": "Overwrite an existing target config when applying. Defaults to false."
-                    }
-                },
-                "required": [],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn provider_switch_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Inspect current provider state or switch the default provider profile for subsequent turns when the user explicitly wants future replies to use another configured provider, profile, or model.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "selector": {
-                        "type": "string",
-                        "description": format!(
-                            "Optional provider selector. Accepts a {} such as `openai-gpt-5`, `gpt-5.1-codex`, or `deepseek`. When omitted, the tool reports current provider state without changing it.",
-                            crate::config::PROVIDER_SELECTOR_HUMAN_SUMMARY
-                        )
-                    }
-                },
-                "required": []
-            }
-        }
-    })
-}
-
-fn external_skills_policy_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Get, set, or reset runtime policy for external skills downloads (enabled flag, approval gate, domain allowlist/blocklist).",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "action": {
-                        "type": "string",
-                        "enum": ["get", "set", "reset"],
-                        "description": "Policy action. Defaults to `get`."
-                    },
-                    "policy_update_approved": {
-                        "type": "boolean",
-                        "description": "Explicit user authorization for policy updates. Required for `set` and `reset`."
-                    },
-                    "enabled": {
-                        "type": "boolean",
-                        "description": "Whether external skills runtime/download is enabled."
-                    },
-                    "require_download_approval": {
-                        "type": "boolean",
-                        "description": "When true, every external skills download requires explicit approval_granted=true."
-                    },
-                    "allowed_domains": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                        "description": "Optional domain allowlist (supports exact domains and wildcard forms like *.example.com). Empty list means allow all domains unless blocked."
-                    },
-                    "blocked_domains": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                        "description": "Optional domain blocklist (supports exact domains and wildcard forms like *.example.com). Blocklist always takes precedence."
-                    }
-                },
-                "required": [],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_fetch_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Resolve and download an external skill artifact from a direct URL, GitHub reference, skills.sh page, clawhub.ai page, or npm package with strict domain policy checks and explicit approval gating.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "reference": {
-                        "type": "string",
-                        "description": "Preferred external skill reference. Supports direct URLs, GitHub refs, skills.sh pages, clawhub.ai pages, and npm packages."
-                    },
-                    "url": {
-                        "type": "string",
-                        "description": "Backward-compatible alias for `reference` when passing a direct URL or ecosystem reference."
-                    },
-                    "approval_granted": {
-                        "type": "boolean",
-                        "description": "Explicit user authorization for this download. Required when require_download_approval=true."
-                    },
-                    "save_as": {
-                        "type": "string",
-                        "description": "Optional output filename (stored under configured file root / external-skills-downloads)."
-                    },
-                    "max_bytes": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 20971520,
-                        "description": "Maximum download size in bytes. Defaults to 5242880 and is capped at 20971520."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_resolve_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Normalize a direct URL, GitHub reference, skills.sh page, ClawHub page, or npm package into a source-aware external skill candidate.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "reference": {
-                        "type": "string",
-                        "description": "External skill reference to normalize."
-                    }
-                },
-                "required": ["reference"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_search_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Search the resolved external-skills inventory for active and shadowed matches.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "type": "string",
-                        "description": "Task phrase, capability phrase, or skill name to rank against discovered skills."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 20,
-                        "description": "Maximum number of ranked matches to return."
-                    }
-                },
-                "required": ["query", "limit"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_recommend_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Recommend the best-fit resolved external skills for an operator goal.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "type": "string",
-                        "description": "Operator goal, task phrase, or workflow description."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 20,
-                        "description": "Maximum number of ranked recommendations to return."
-                    }
-                },
-                "required": ["query", "limit"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_source_search_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Search preferred external skill ecosystems and return normalized source-aware candidates ranked by source priority.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "type": "string",
-                        "description": "Search query or external skill reference."
-                    },
-                    "max_results": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 20,
-                        "description": "Maximum number of normalized candidates to return."
-                    },
-                    "sources": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                        "description": "Optional source filter list. Supported values: skills_sh, clawhub, github, npm."
-                    }
-                },
-                "required": ["query"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_inspect_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Read metadata and a short preview for a resolved external skill across managed, user, and project scopes.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "skill_id": {
-                        "type": "string",
-                        "description": "Resolved external skill identifier."
-                    }
-                },
-                "required": ["skill_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_install_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Install a managed external skill from a local directory, local .tgz/.tar.gz/.zip archive, or a first-party bundled skill id.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "path": {
-                        "type": "string",
-                        "description": "Path to a local directory containing SKILL.md or a local .tgz/.tar.gz/.zip archive."
-                    },
-                    "bundled_skill_id": {
-                        "type": "string",
-                        "description": "Optional first-party bundled skill identifier, for example `browser-companion-preview`."
-                    },
-                    "skill_id": {
-                        "type": "string",
-                        "description": "Optional explicit managed skill id override."
-                    },
-                    "source_skill_id": {
-                        "type": "string",
-                        "description": "Optional source skill selector when the input archive or directory contains multiple SKILL.md roots."
-                    },
-                    "security_decision": {
-                        "type": "string",
-                        "enum": ["approve_once", "deny"],
-                        "description": "Optional one-time security override after a risky install was scanned and returned needs_approval."
-                    },
-                    "replace": {
-                        "type": "boolean",
-                        "description": "Replace an existing installed skill with the same id. Defaults to false."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_invoke_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Load a resolved external skill's SKILL.md instructions into the conversation loop across managed, user, and project scopes.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "skill_id": {
-                        "type": "string",
-                        "description": "Resolved external skill identifier."
-                    }
-                },
-                "required": ["skill_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_list_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "List resolved external skills available for invocation across managed, user, and project scopes.",
-            "parameters": {
-                "type": "object",
-                "properties": {},
-                "required": [],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn external_skills_remove_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": "Remove an installed external skill from the managed runtime.",
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "skill_id": {
-                        "type": "string",
-                        "description": "Managed external skill identifier."
-                    }
-                },
-                "required": ["skill_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
 fn file_read_definition(descriptor: &ToolDescriptor) -> Value {
     json!({
         "type": "function",
@@ -3634,6 +2530,25 @@ fn memory_get_definition(descriptor: &ToolDescriptor) -> Value {
                 "additionalProperties": false
             }
         }
+    })
+}
+
+fn exact_edit_block_definition() -> Value {
+    json!({
+        "type": "object",
+        "properties": {
+            "old_text": {
+                "type": "string",
+                "minLength": 1,
+                "description": "Exact text for one targeted replacement. It must match uniquely in the original file and must not overlap any other edit block."
+            },
+            "new_text": {
+                "type": "string",
+                "description": "Replacement text for this targeted edit block."
+            }
+        },
+        "required": ["old_text", "new_text"],
+        "additionalProperties": false
     })
 }
 
@@ -3863,685 +2778,6 @@ fn bash_exec_definition(descriptor: &ToolDescriptor) -> Value {
     })
 }
 
-fn approval_request_resolve_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "approval_request_id": {
-                        "type": "string",
-                        "description": "Visible approval request identifier to resolve."
-                    },
-                    "decision": {
-                        "type": "string",
-                        "enum": ["approve_once", "approve_always", "deny"],
-                        "description": "Operator decision for the pending approval request."
-                    },
-                    "session_consent_mode": {
-                        "type": "string",
-                        "enum": ["auto", "full"],
-                        "description": "Optional session consent mode to persist when approve_once wins the request."
-                    }
-                },
-                "required": ["approval_request_id", "decision"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn approval_request_status_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "approval_request_id": {
-                        "type": "string",
-                        "description": "Visible approval request identifier to inspect in detail."
-                    }
-                },
-                "required": ["approval_request_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn approval_requests_list_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Optional visible session identifier to scope approval requests to one session."
-                    },
-                    "status": {
-                        "type": "string",
-                        "enum": ["pending", "approved", "executing", "executed", "denied", "expired", "cancelled"],
-                        "description": "Optional approval request status filter."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 100,
-                        "description": "Maximum visible approval requests to return after filtering."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn sessions_list_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 200,
-                        "description": "Maximum visible sessions to return after filtering."
-                    },
-                    "offset": {
-                        "type": "integer",
-                        "minimum": 0,
-                        "description": "Number of matching visible sessions to skip before applying limit."
-                    },
-                    "state": {
-                        "type": "string",
-                        "enum": ["ready", "running", "completed", "failed", "timed_out"],
-                        "description": "Optional lifecycle state filter."
-                    },
-                    "kind": {
-                        "type": "string",
-                        "enum": ["root", "delegate_child"],
-                        "description": "Optional session kind filter."
-                    },
-                    "parent_session_id": {
-                        "type": "string",
-                        "description": "Optional direct parent session filter."
-                    },
-                    "overdue_only": {
-                        "type": "boolean",
-                        "description": "When true, only return async delegate children whose lifecycle staleness is overdue."
-                    },
-                    "include_archived": {
-                        "type": "boolean",
-                        "description": "When true, include archived visible sessions in the returned list."
-                    },
-                    "include_delegate_lifecycle": {
-                        "type": "boolean",
-                        "description": "When true, include normalized delegate lifecycle metadata for returned sessions."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn sessions_history_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible session identifier to inspect."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 200,
-                        "description": "Maximum transcript entries to return."
-                    }
-                },
-                "required": ["session_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_events_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible session identifier to inspect."
-                    },
-                    "after_id": {
-                        "type": "integer",
-                        "minimum": 0,
-                        "description": "Optional event id cursor; when present only newer events are returned."
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 200,
-                        "description": "Maximum event rows to return."
-                    }
-                },
-                "required": ["session_id"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_search_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "query": {
-                        "type": "string",
-                        "description": "Natural-language search query over visible canonical session history."
-                    },
-                    "session_id": {
-                        "type": "string",
-                        "description": "Optional visible session id to narrow the search scope."
-                    },
-                    "max_results": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 20,
-                        "description": "Optional maximum number of ranked hits to return. Defaults to 5."
-                    },
-                    "include_archived": {
-                        "type": "boolean",
-                        "description": "Include archived visible sessions when true. Defaults to false."
-                    },
-                    "include_turns": {
-                        "type": "boolean",
-                        "description": "Include transcript turn matches. Defaults to true."
-                    },
-                    "include_events": {
-                        "type": "boolean",
-                        "description": "Include session event matches. Defaults to true."
-                    }
-                },
-                "required": ["query"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_status_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible session identifier to inspect."
-                    },
-                    "session_ids": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "minItems": 1,
-                        "description": "Visible session identifiers to inspect in one request."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_tool_runtime_narrowing_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "browser": {
-                "type": "object",
-                "properties": {
-                    "max_sessions": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional upper bound for browser session count."
-                    },
-                    "max_links": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional upper bound for extracted browser links."
-                    },
-                    "max_text_chars": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional upper bound for extracted browser text characters."
-                    }
-                },
-                "additionalProperties": false
-            },
-            "web_fetch": {
-                "type": "object",
-                "properties": {
-                    "allow_private_hosts": {
-                        "type": "boolean",
-                        "description": "Optional narrowing for private-host access. Use false to deny private hosts."
-                    },
-                    "enforce_allowed_domains": {
-                        "type": "boolean",
-                        "description": "When true, enforce the provided allowed_domains list even when it is empty."
-                    },
-                    "allowed_domains": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "description": "Optional allowlist intersection for web.fetch."
-                    },
-                    "blocked_domains": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "description": "Optional additional blocked domains for web.fetch."
-                    },
-                    "timeout_seconds": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional maximum web.fetch timeout."
-                    },
-                    "max_bytes": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional maximum web.fetch response size in bytes."
-                    },
-                    "max_redirects": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "description": "Optional maximum web.fetch redirect count."
-                    }
-                },
-                "additionalProperties": false
-            }
-        },
-        "additionalProperties": false
-    })
-}
-
-fn session_tool_policy_status_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Optional visible session identifier to inspect. Defaults to the current session."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_tool_policy_set_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Optional visible session identifier to update. Defaults to the current session."
-                    },
-                    "tool_ids": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "description": "Optional replacement visible tool id set. Use an empty array to clear the session-specific tool surface restriction."
-                    },
-                    "runtime_narrowing": session_tool_runtime_narrowing_schema()
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_tool_policy_clear_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Optional visible session identifier to clear. Defaults to the current session."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_recover_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible delegate child session identifier to recover."
-                    },
-                    "session_ids": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "minItems": 1,
-                        "description": "Visible delegate child session identifiers to recover in one request."
-                    },
-                    "dry_run": {
-                        "type": "boolean",
-                        "description": "When true, preview which targets are recoverable without mutating state."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_archive_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible terminal session identifier to archive."
-                    },
-                    "session_ids": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "minItems": 1,
-                        "description": "Visible terminal session identifiers to archive in one request."
-                    },
-                    "dry_run": {
-                        "type": "boolean",
-                        "description": "When true, preview which targets are archivable without mutating state."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_cancel_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible async delegate child session identifier to cancel."
-                    },
-                    "session_ids": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "minItems": 1,
-                        "description": "Visible async delegate child session identifiers to cancel in one request."
-                    },
-                    "dry_run": {
-                        "type": "boolean",
-                        "description": "When true, preview which targets are cancellable without mutating state."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_continue_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible delegate child session identifier to continue."
-                    },
-                    "input": {
-                        "type": "string",
-                        "description": "Follow-up user input to execute inside the target child session."
-                    },
-                    "timeout_seconds": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 600,
-                        "description": "Optional bounded timeout override for the continued child turn."
-                    }
-                },
-                "required": ["session_id", "input"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn session_wait_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Visible session identifier to wait on."
-                    },
-                    "session_ids": {
-                        "type": "array",
-                        "items": {
-                            "type": "string"
-                        },
-                        "minItems": 1,
-                        "description": "Visible session identifiers to wait on in one request."
-                    },
-                    "after_id": {
-                        "type": "integer",
-                        "description": "Optional event cursor. When present, the response also returns session events with id greater than this value."
-                    },
-                    "timeout_ms": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 30000,
-                        "description": "Bounded wait timeout in milliseconds."
-                    }
-                },
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn sessions_send_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "session_id": {
-                        "type": "string",
-                        "description": "Known channel-backed root session identifier to receive the outbound text message (for example Telegram, Feishu, or Matrix)."
-                    },
-                    "text": {
-                        "type": "string",
-                        "description": "Outbound plain-text message content."
-                    }
-                },
-                "required": ["session_id", "text"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn delegate_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "task": {
-                        "type": "string",
-                        "description": "Focused subtask to run in a child session."
-                    },
-                    "label": {
-                        "type": "string",
-                        "description": "Optional human-readable label for the child session."
-                    },
-                    "profile": {
-                        "type": "string",
-                        "enum": ["research", "plan", "verify"],
-                        "description": "Optional builtin child profile preset. `research`, `plan`, and `verify` apply bounded delegate role defaults."
-                    },
-                    "isolation": {
-                        "type": "string",
-                        "enum": ["shared", "worktree"],
-                        "description": "Optional child workspace isolation mode. `shared` reuses the current workspace root. `worktree` is reserved for a dedicated git worktree-backed child root and currently returns a not-supported error until that runtime lane lands."
-                    },
-                    "timeout_seconds": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 600,
-                        "description": "Optional timeout for the delegated task."
-                    }
-                },
-                "required": ["task"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
-fn delegate_async_definition(descriptor: &ToolDescriptor) -> Value {
-    json!({
-        "type": "function",
-        "function": {
-            "name": descriptor.provider_name,
-            "description": descriptor.description,
-            "parameters": {
-                "type": "object",
-                "properties": {
-                    "task": {
-                        "type": "string",
-                        "description": "Focused subtask to run in a background child session."
-                    },
-                    "label": {
-                        "type": "string",
-                        "description": "Optional human-readable label for the child session."
-                    },
-                    "profile": {
-                        "type": "string",
-                        "enum": ["research", "plan", "verify"],
-                        "description": "Optional builtin child profile preset. `research`, `plan`, and `verify` apply bounded delegate role defaults."
-                    },
-                    "isolation": {
-                        "type": "string",
-                        "enum": ["shared", "worktree"],
-                        "description": "Optional child workspace isolation mode. `shared` reuses the current workspace root. `worktree` is reserved for a dedicated git worktree-backed child root and currently returns a not-supported error until that runtime lane lands."
-                    },
-                    "timeout_seconds": {
-                        "type": "integer",
-                        "minimum": 1,
-                        "maximum": 600,
-                        "description": "Optional timeout for the delegated task."
-                    }
-                },
-                "required": ["task"],
-                "additionalProperties": false
-            }
-        }
-    })
-}
-
 #[cfg(feature = "feishu-integration")]
 fn push_feishu_tool_descriptor(
     descriptors: &mut Vec<ToolDescriptor>,
@@ -4584,879 +2820,6 @@ fn feishu_definition(descriptor: &ToolDescriptor) -> Value {
     })
 }
 
-fn tool_argument_hint(name: &str) -> &'static str {
-    if let Some(argument_hint) = super::tool_surface::direct_tool_argument_hint(name) {
-        return argument_hint;
-    }
-
-    match name {
-        "feishu.bitable.app.create" => {
-            "account_id?:string,open_id?:string,name:string,folder_token?:string"
-        }
-        "feishu.bitable.app.get" => "account_id?:string,open_id?:string,app_token:string",
-        "feishu.bitable.app.list" => {
-            "account_id?:string,open_id?:string,folder_token?:string,page_size?:integer,page_token?:string"
-        }
-        "feishu.bitable.app.patch" => {
-            "account_id?:string,open_id?:string,app_token:string,name?:string,is_advanced?:boolean"
-        }
-        "feishu.bitable.app.copy" => {
-            "account_id?:string,open_id?:string,app_token:string,name:string,folder_token?:string"
-        }
-        "feishu.bitable.list" => {
-            "account_id?:string,open_id?:string,app_token:string,page_size?:integer,page_token?:string"
-        }
-        "feishu.bitable.table.create" => {
-            "account_id?:string,open_id?:string,app_token:string,name:string,default_view_name?:string,fields?:array"
-        }
-        "feishu.bitable.table.patch" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,name:string"
-        }
-        "feishu.bitable.table.batch_create" => {
-            "account_id?:string,open_id?:string,app_token:string,tables:array"
-        }
-        "feishu.bitable.record.create" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,fields:object"
-        }
-        "feishu.bitable.record.update" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,record_id:string,fields:object"
-        }
-        "feishu.bitable.record.delete" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,record_id:string"
-        }
-        "feishu.bitable.record.batch_create" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,records:array"
-        }
-        "feishu.bitable.record.batch_update" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,records:array"
-        }
-        "feishu.bitable.record.batch_delete" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,records:array"
-        }
-        "feishu.bitable.field.create" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,field_name:string,type:integer,property?:object"
-        }
-        "feishu.bitable.field.list" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,view_id?:string,page_size?:integer,page_token?:string"
-        }
-        "feishu.bitable.field.update" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,field_id:string,field_name:string,type:integer,property?:object"
-        }
-        "feishu.bitable.field.delete" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,field_id:string"
-        }
-        "feishu.bitable.view.create" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,view_name:string,view_type?:string"
-        }
-        "feishu.bitable.view.get" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,view_id:string"
-        }
-        "feishu.bitable.view.list" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,page_size?:integer,page_token?:string"
-        }
-        "feishu.bitable.view.patch" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,view_id:string,view_name:string"
-        }
-        "feishu.bitable.record.search" => {
-            "account_id?:string,open_id?:string,app_token:string,table_id:string,view_id?:string,filter?:object,sort?:array,field_names?:string[],automatic_fields?:boolean,page_size?:integer,page_token?:string"
-        }
-        "feishu.calendar.freebusy" => {
-            "account_id?:string,open_id?:string,time_min:string,time_max:string,user_id?:string,room_id?:string"
-        }
-        "feishu.calendar.list" => {
-            "account_id?:string,open_id?:string,primary?:boolean,page_size?:integer,page_token?:string,sync_token?:string"
-        }
-        "feishu.calendar.primary.get" => "account_id?:string,open_id?:string,user_id_type?:string",
-        "feishu.card.update" => {
-            "account_id?:string,callback_token?:string,card?:object,markdown?:string,shared?:boolean,open_ids?:string[]"
-        }
-        "feishu.doc.append" => {
-            "account_id?:string,open_id?:string,url:string,content?:string,content_path?:string,content_type?:string"
-        }
-        "feishu.doc.create" => {
-            "account_id?:string,open_id?:string,title?:string,folder_token?:string,content?:string,content_path?:string,content_type?:string"
-        }
-        "feishu.doc.read" => "account_id?:string,open_id?:string,url:string,lang?:integer",
-        "feishu.messages.get" => "account_id?:string,open_id?:string,message_id:string",
-        "feishu.messages.history" => {
-            "account_id?:string,open_id?:string,container_id?:string,container_id_type?:string,page_size?:integer,page_token?:string"
-        }
-        #[cfg(feature = "tool-file")]
-        "feishu.messages.resource.get" => {
-            "account_id?:string,open_id?:string,message_id?:string,file_key?:string,type?:string,save_as?:string"
-        }
-        "feishu.messages.reply" => {
-            "account_id?:string,open_id?:string,message_id:string,text?:string,post?:object,image_key?:string,file_key?:string,card?:object,markdown?:string"
-        }
-        "feishu.messages.search" => {
-            "account_id?:string,open_id?:string,query:string,page_size?:integer,page_token?:string"
-        }
-        "feishu.messages.send" => {
-            "account_id?:string,open_id?:string,receive_id:string,receive_id_type?:string,text?:string,post?:object,image_key?:string,file_key?:string,card?:object,markdown?:string"
-        }
-        "feishu.whoami" => "account_id?:string,open_id?:string",
-        "tool.search" => "query?:string,exact_tool_id?:string,limit?:integer",
-        "tool.invoke" => "tool_id:string,lease:string,arguments:object",
-        "read" => {
-            "path?:string,offset?:integer,limit?:integer,max_bytes?:integer,query?:string,pattern?:string,root?:string,glob?:string,max_results?:integer,max_bytes_per_file?:integer,case_sensitive?:boolean,include_directories?:boolean"
-        }
-        "write" => {
-            "path:string,content?:string,create_dirs?:boolean,overwrite?:boolean,edits?:array,old_string?:string,new_string?:string,replace_all?:boolean"
-        }
-        "exec" => "command?:string,script?:string,args?:string[],timeout_ms?:integer,cwd?:string",
-        "web" => {
-            "url?:string,mode?:string,max_bytes?:integer,query?:string,provider?:string,max_results?:integer"
-        }
-        "browser" => {
-            "url?:string,max_bytes?:integer,session_id?:string,mode?:string,selector?:string,limit?:integer,link_id?:integer"
-        }
-        "memory" => "query?:string,max_results?:integer,path?:string,from?:integer,lines?:integer",
-        "config.import" => {
-            "input_path?:string,output_path?:string,mode?:string,source?:string,source_id?:string,primary_source_id?:string,safe_profile_merge?:boolean,apply_external_skills_plan?:boolean,force?:boolean"
-        }
-        "external_skills.fetch" => {
-            "reference?:string,url?:string,approval_granted?:boolean,save_as?:string,max_bytes?:integer"
-        }
-        "external_skills.resolve" => "reference:string",
-        "external_skills.search" => "query:string,limit:integer",
-        "external_skills.recommend" => "query:string,limit:integer",
-        "external_skills.source_search" => "query:string,max_results?:integer,sources?:string[]",
-        "external_skills.inspect" => "skill_id:string",
-        "external_skills.install" => {
-            "path?:string,bundled_skill_id?:string,skill_id?:string,source_skill_id?:string,security_decision?:string,replace?:boolean"
-        }
-        "external_skills.invoke" => "skill_id:string",
-        "external_skills.list" => "",
-        "external_skills.policy" => {
-            "action?:string,enabled?:boolean,allowed_domains?:string[],blocked_domains?:string[]"
-        }
-        "external_skills.remove" => "skill_id:string",
-        "browser.companion.session.start" => "url:string",
-        "browser.companion.navigate" => "session_id:string,url:string",
-        "browser.companion.snapshot" => "session_id:string,mode?:string",
-        "browser.companion.wait" => "session_id:string,condition?:string,timeout_ms?:integer",
-        "browser.companion.session.stop" => "session_id:string",
-        "browser.companion.click" => "session_id:string,selector:string",
-        "browser.companion.type" => "session_id:string,selector:string,text:string",
-        "http.request" => {
-            "url:string,method?:string,headers?:object,body?:string,content_type?:string,max_bytes?:integer"
-        }
-        "file.read" => "path:string,offset?:integer,limit?:integer,max_bytes?:integer",
-        "glob.search" => {
-            "pattern:string,root?:string,max_results?:integer,include_directories?:boolean"
-        }
-        "content.search" => {
-            "query:string,root?:string,glob?:string,max_results?:integer,max_bytes_per_file?:integer,case_sensitive?:boolean"
-        }
-        "memory_search" => "query:string,max_results?:integer",
-        "memory_get" => "path:string,from?:integer,lines?:integer",
-        "file.write" => "path:string,content:string,create_dirs?:boolean,overwrite?:boolean",
-        "file.edit" => {
-            "path:string,edits?:array,old_string?:string,new_string?:string,replace_all?:boolean"
-        }
-        "shell.exec" => "command:string,args?:string[],timeout_ms?:integer,cwd?:string",
-        "bash.exec" => "command:string,cwd?:string,timeout_ms?:integer",
-        "provider.switch" => "selector?:string",
-        "delegate" | "delegate_async" => {
-            "task:string,label?:string,profile?:string,isolation?:string,timeout_seconds?:integer"
-        }
-        "session_archive" | "session_cancel" | "session_events" | "session_recover"
-        | "session_status" | "session_wait" | "sessions_history" => "session_id:string",
-        "session_continue" => "session_id:string,input:string,timeout_seconds?:integer",
-        "sessions_list" => "limit?:integer,offset?:integer,state?:string",
-        "sessions_send" => "session_id:string,text:string",
-        "web.search" => "query:string,provider?:string,max_results?:integer",
-        _ => "",
-    }
-}
-
-fn tool_search_hint(name: &str, fallback: &'static str) -> &'static str {
-    if let Some(search_hint) = super::tool_surface::direct_tool_search_hint(name) {
-        return search_hint;
-    }
-
-    match name {
-        "tool.search" => {
-            "discover a specialized tool when the visible direct tools do not fit, or refresh a known tool card"
-        }
-        "tool.invoke" => "run a discovered specialized tool with the lease returned by tool.search",
-        "http.request" => {
-            "send a bounded http request, inspect status and headers, fetch text or binary responses"
-        }
-        "file.read" => {
-            "read a workspace file, inspect file contents, or page through a file window"
-        }
-        "glob.search" => {
-            "find workspace files by glob pattern, list files in a directory, browse folder contents, search repo paths, match files under a root"
-        }
-        "content.search" => {
-            "search workspace file contents, find text in repo files, grep text in the project"
-        }
-        "file.write" => {
-            "write a workspace file, save file content, create or overwrite a repo file"
-        }
-        "file.edit" => {
-            "edit a workspace file, patch file content, or apply exact replacement blocks in a repo file"
-        }
-        "shell.exec" => {
-            "run a shell command, execute a terminal command, bash, zsh, powershell, cli"
-        }
-        "web.fetch" => "fetch a web page, download page text, inspect http content from a url",
-        "web.search" => "search the web, look up web results, find information online",
-        "memory_search" => {
-            "search durable workspace memory, recall prior notes, query stored memory"
-        }
-        "memory_get" => "read a memory note by path, inspect saved durable memory content",
-        "provider.switch" => "switch model provider, change runtime provider selection",
-        _ => fallback,
-    }
-}
-
-fn tool_parameter_types(name: &str) -> &'static [(&'static str, &'static str)] {
-    if let Some(parameter_types) = super::tool_surface::direct_tool_parameter_types(name) {
-        return parameter_types;
-    }
-
-    match name {
-        "feishu.bitable.app.create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("name", "string"),
-            ("folder_token", "string"),
-        ],
-        "feishu.bitable.app.get" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-        ],
-        "feishu.bitable.app.list" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("folder_token", "string"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        "feishu.bitable.app.patch" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("name", "string"),
-            ("is_advanced", "boolean"),
-        ],
-        "feishu.bitable.app.copy" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("name", "string"),
-            ("folder_token", "string"),
-        ],
-        "feishu.bitable.list" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        "feishu.bitable.table.create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("name", "string"),
-            ("default_view_name", "string"),
-            ("fields", "array"),
-        ],
-        "feishu.bitable.table.patch" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("name", "string"),
-        ],
-        "feishu.bitable.table.batch_create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("tables", "array"),
-        ],
-        "feishu.bitable.record.create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("fields", "object"),
-        ],
-        "feishu.bitable.record.update" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("record_id", "string"),
-            ("fields", "object"),
-        ],
-        "feishu.bitable.record.delete" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("record_id", "string"),
-        ],
-        "feishu.bitable.record.batch_create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("records", "array"),
-        ],
-        "feishu.bitable.record.batch_update" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("records", "array"),
-        ],
-        "feishu.bitable.record.batch_delete" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("records", "array"),
-        ],
-        "feishu.bitable.field.create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("field_name", "string"),
-            ("type", "integer"),
-            ("property", "object"),
-        ],
-        "feishu.bitable.field.list" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("view_id", "string"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        "feishu.bitable.field.update" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("field_id", "string"),
-            ("field_name", "string"),
-            ("type", "integer"),
-            ("property", "object"),
-        ],
-        "feishu.bitable.field.delete" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("field_id", "string"),
-        ],
-        "feishu.bitable.view.create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("view_name", "string"),
-            ("view_type", "string"),
-        ],
-        "feishu.bitable.view.get" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("view_id", "string"),
-        ],
-        "feishu.bitable.view.list" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        "feishu.bitable.view.patch" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("view_id", "string"),
-            ("view_name", "string"),
-        ],
-        "feishu.bitable.record.search" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("app_token", "string"),
-            ("table_id", "string"),
-            ("view_id", "string"),
-            ("filter", "object"),
-            ("sort", "array"),
-            ("field_names", "array"),
-            ("automatic_fields", "boolean"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        "feishu.calendar.freebusy" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("time_min", "string"),
-            ("time_max", "string"),
-            ("user_id", "string"),
-            ("room_id", "string"),
-        ],
-        "feishu.calendar.list" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("primary", "boolean"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-            ("sync_token", "string"),
-        ],
-        "feishu.calendar.primary.get" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("user_id_type", "string"),
-        ],
-        "feishu.card.update" => &[
-            ("account_id", "string"),
-            ("callback_token", "string"),
-            ("card", "object"),
-            ("markdown", "string"),
-            ("shared", "boolean"),
-            ("open_ids", "array"),
-        ],
-        "feishu.doc.append" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("url", "string"),
-            ("content", "string"),
-            ("content_path", "string"),
-            ("content_type", "string"),
-        ],
-        "feishu.doc.create" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("title", "string"),
-            ("folder_token", "string"),
-            ("content", "string"),
-            ("content_path", "string"),
-            ("content_type", "string"),
-        ],
-        "feishu.doc.read" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("url", "string"),
-            ("lang", "integer"),
-        ],
-        "feishu.messages.get" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("message_id", "string"),
-        ],
-        "feishu.messages.history" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("container_id", "string"),
-            ("container_id_type", "string"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        #[cfg(feature = "tool-file")]
-        "feishu.messages.resource.get" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("message_id", "string"),
-            ("file_key", "string"),
-            ("type", "string"),
-            ("save_as", "string"),
-        ],
-        "feishu.messages.reply" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("message_id", "string"),
-            ("text", "string"),
-            ("post", "object"),
-            ("image_key", "string"),
-            ("file_key", "string"),
-            ("card", "object"),
-            ("markdown", "string"),
-        ],
-        "feishu.messages.search" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("query", "string"),
-            ("page_size", "integer"),
-            ("page_token", "string"),
-        ],
-        "feishu.messages.send" => &[
-            ("account_id", "string"),
-            ("open_id", "string"),
-            ("receive_id", "string"),
-            ("receive_id_type", "string"),
-            ("text", "string"),
-            ("post", "object"),
-            ("image_key", "string"),
-            ("file_key", "string"),
-            ("card", "object"),
-            ("markdown", "string"),
-        ],
-        "feishu.whoami" => &[("account_id", "string"), ("open_id", "string")],
-        "tool.search" => &[
-            ("query", "string"),
-            ("exact_tool_id", "string"),
-            ("limit", "integer"),
-        ],
-        "tool.invoke" => &[
-            ("tool_id", "string"),
-            ("lease", "string"),
-            ("arguments", "object"),
-        ],
-        "config.import" => &[
-            ("input_path", "string"),
-            ("output_path", "string"),
-            ("mode", "string"),
-            ("source", "string"),
-            ("source_id", "string"),
-            ("primary_source_id", "string"),
-            ("safe_profile_merge", "boolean"),
-            ("apply_external_skills_plan", "boolean"),
-            ("force", "boolean"),
-        ],
-        "external_skills.fetch" => &[
-            ("reference", "string"),
-            ("url", "string"),
-            ("approval_granted", "boolean"),
-            ("save_as", "string"),
-            ("max_bytes", "integer"),
-        ],
-        "external_skills.resolve" => &[("reference", "string")],
-        "external_skills.search" => &[("query", "string"), ("limit", "integer")],
-        "external_skills.recommend" => &[("query", "string"), ("limit", "integer")],
-        "external_skills.source_search" => &[
-            ("query", "string"),
-            ("max_results", "integer"),
-            ("sources", "array"),
-        ],
-        "external_skills.inspect" | "external_skills.invoke" | "external_skills.remove" => {
-            &[("skill_id", "string")]
-        }
-        "external_skills.install" => &[
-            ("path", "string"),
-            ("bundled_skill_id", "string"),
-            ("skill_id", "string"),
-            ("source_skill_id", "string"),
-            ("security_decision", "string"),
-            ("replace", "boolean"),
-        ],
-        "external_skills.list" => &[],
-        "browser.companion.session.start" => &[("url", "string")],
-        "browser.companion.navigate" => &[("session_id", "string"), ("url", "string")],
-        "browser.companion.snapshot" => &[("session_id", "string"), ("mode", "string")],
-        "browser.companion.wait" => &[
-            ("session_id", "string"),
-            ("condition", "string"),
-            ("timeout_ms", "integer"),
-        ],
-        "browser.companion.session.stop" => &[("session_id", "string")],
-        "browser.companion.click" => &[("session_id", "string"), ("selector", "string")],
-        "browser.companion.type" => &[
-            ("session_id", "string"),
-            ("selector", "string"),
-            ("text", "string"),
-        ],
-        "http.request" => &[
-            ("url", "string"),
-            ("method", "string"),
-            ("headers", "object"),
-            ("body", "string"),
-            ("content_type", "string"),
-            ("max_bytes", "integer"),
-        ],
-        "external_skills.policy" => &[
-            ("action", "string"),
-            ("enabled", "boolean"),
-            ("allowed_domains", "array"),
-            ("blocked_domains", "array"),
-        ],
-        "file.read" => &[
-            ("path", "string"),
-            ("offset", "integer"),
-            ("limit", "integer"),
-            ("max_bytes", "integer"),
-        ],
-        "glob.search" => &[
-            ("pattern", "string"),
-            ("root", "string"),
-            ("max_results", "integer"),
-            ("include_directories", "boolean"),
-        ],
-        "content.search" => &[
-            ("query", "string"),
-            ("root", "string"),
-            ("glob", "string"),
-            ("max_results", "integer"),
-            ("max_bytes_per_file", "integer"),
-            ("case_sensitive", "boolean"),
-        ],
-        "memory_search" => &[("query", "string"), ("max_results", "integer")],
-        "memory_get" => &[
-            ("path", "string"),
-            ("from", "integer"),
-            ("lines", "integer"),
-        ],
-        "file.write" => &[
-            ("path", "string"),
-            ("content", "string"),
-            ("create_dirs", "boolean"),
-            ("overwrite", "boolean"),
-        ],
-        "file.edit" => &[
-            ("path", "string"),
-            ("edits", "array"),
-            ("old_string", "string"),
-            ("new_string", "string"),
-            ("replace_all", "boolean"),
-        ],
-        "shell.exec" => &[
-            ("command", "string"),
-            ("args", "array"),
-            ("timeout_ms", "integer"),
-            ("cwd", "string"),
-        ],
-        "bash.exec" => &[
-            ("command", "string"),
-            ("cwd", "string"),
-            ("timeout_ms", "integer"),
-        ],
-        "provider.switch" => &[("selector", "string")],
-        "delegate" | "delegate_async" => &[
-            ("task", "string"),
-            ("label", "string"),
-            ("profile", "string"),
-            ("isolation", "string"),
-            ("timeout_seconds", "integer"),
-        ],
-        "session_continue" => &[
-            ("session_id", "string"),
-            ("input", "string"),
-            ("timeout_seconds", "integer"),
-        ],
-        "session_tool_policy_status" | "session_tool_policy_clear" => &[("session_id", "string")],
-        "session_tool_policy_set" => &[
-            ("session_id", "string"),
-            ("tool_ids", "array"),
-            ("runtime_narrowing", "object"),
-        ],
-        "session_archive" | "session_cancel" | "session_events" | "session_recover"
-        | "session_status" | "session_wait" | "sessions_history" => &[("session_id", "string")],
-        "sessions_list" => &[
-            ("limit", "integer"),
-            ("offset", "integer"),
-            ("state", "string"),
-        ],
-        "session_search" => &[
-            ("query", "string"),
-            ("session_id", "string"),
-            ("max_results", "integer"),
-            ("include_archived", "boolean"),
-            ("include_turns", "boolean"),
-            ("include_events", "boolean"),
-        ],
-        "sessions_send" => &[("session_id", "string"), ("text", "string")],
-        "web.search" => &[
-            ("query", "string"),
-            ("provider", "string"),
-            ("max_results", "integer"),
-        ],
-        _ => &[],
-    }
-}
-
-fn tool_required_fields(name: &str) -> &'static [&'static str] {
-    if let Some(required_fields) = super::tool_surface::direct_tool_required_fields(name) {
-        return required_fields;
-    }
-
-    match name {
-        "feishu.bitable.app.create" => &["name"],
-        "feishu.bitable.app.get" => &["app_token"],
-        "feishu.bitable.app.list" => &[],
-        "feishu.bitable.app.patch" => &["app_token"],
-        "feishu.bitable.app.copy" => &["app_token", "name"],
-        "feishu.bitable.list" => &["app_token"],
-        "feishu.bitable.table.create" => &["app_token", "name"],
-        "feishu.bitable.table.patch" => &["app_token", "table_id", "name"],
-        "feishu.bitable.table.batch_create" => &["app_token", "tables"],
-        "feishu.bitable.record.create" => &["app_token", "table_id", "fields"],
-        "feishu.bitable.record.update" => &["app_token", "table_id", "record_id", "fields"],
-        "feishu.bitable.record.delete" => &["app_token", "table_id", "record_id"],
-        "feishu.bitable.record.batch_create"
-        | "feishu.bitable.record.batch_update"
-        | "feishu.bitable.record.batch_delete" => &["app_token", "table_id", "records"],
-        "feishu.bitable.field.create" => &["app_token", "table_id", "field_name", "type"],
-        "feishu.bitable.field.list" => &["app_token", "table_id"],
-        "feishu.bitable.field.update" => {
-            &["app_token", "table_id", "field_id", "field_name", "type"]
-        }
-        "feishu.bitable.field.delete" => &["app_token", "table_id", "field_id"],
-        "feishu.bitable.view.create" => &["app_token", "table_id", "view_name"],
-        "feishu.bitable.view.get" => &["app_token", "table_id", "view_id"],
-        "feishu.bitable.view.list" => &["app_token", "table_id"],
-        "feishu.bitable.view.patch" => &["app_token", "table_id", "view_id", "view_name"],
-        "feishu.bitable.record.search" => &["app_token", "table_id"],
-        "feishu.calendar.freebusy" => &["time_min", "time_max"],
-        "feishu.doc.append" | "feishu.doc.read" => &["url"],
-        "feishu.messages.get" => &["message_id"],
-        "feishu.messages.reply" => &["message_id"],
-        "feishu.messages.search" => &["query"],
-        "feishu.messages.send" => &["receive_id"],
-        "tool.search" => &[],
-        "tool.invoke" => &["tool_id", "lease", "arguments"],
-        "external_skills.fetch" => &[],
-        "external_skills.resolve" => &["reference"],
-        "external_skills.search" => &["query", "limit"],
-        "external_skills.recommend" => &["query", "limit"],
-        "external_skills.source_search" => &["query"],
-        "external_skills.inspect" | "external_skills.invoke" | "external_skills.remove" => {
-            &["skill_id"]
-        }
-        // Grouped requirements are the source of truth for this tool's anyOf shape.
-        "external_skills.install" => &[],
-        "browser.companion.session.start" => &["url"],
-        "browser.companion.navigate" => &["session_id", "url"],
-        "browser.companion.snapshot"
-        | "browser.companion.wait"
-        | "browser.companion.session.stop" => &["session_id"],
-        "browser.companion.click" => &["session_id", "selector"],
-        "browser.companion.type" => &["session_id", "selector", "text"],
-        "http.request" => &["url"],
-        "file.read" => &["path"],
-        "glob.search" => &["pattern"],
-        "content.search" => &["query"],
-        "memory_search" => &["query"],
-        "memory_get" => &["path"],
-        "file.write" => &["path", "content"],
-        "file.edit" => &["path"],
-        "shell.exec" => &["command"],
-        "bash.exec" => &["command"],
-        "delegate" | "delegate_async" => &["task"],
-        "session_tool_policy_status" | "session_tool_policy_clear" => &[],
-        "session_tool_policy_set" => &[],
-        "session_archive" | "session_cancel" | "session_events" | "session_recover"
-        | "session_status" | "session_wait" | "sessions_history" => &["session_id"],
-        "session_continue" => &["session_id", "input"],
-        "sessions_send" => &["session_id", "text"],
-        "web.search" => &["query"],
-        _ => &[],
-    }
-}
-
-fn tool_tags(name: &str) -> &'static [&'static str] {
-    if let Some(tags) = super::tool_surface::direct_tool_tags(name) {
-        return tags;
-    }
-
-    match name {
-        "feishu.bitable.app.get" | "feishu.bitable.app.list" => {
-            &["feishu", "bitable", "app", "read"]
-        }
-        "feishu.bitable.app.create" | "feishu.bitable.app.patch" | "feishu.bitable.app.copy" => {
-            &["feishu", "bitable", "app", "write"]
-        }
-        "feishu.bitable.list" | "feishu.bitable.record.search" => &["feishu", "bitable", "read"],
-        "feishu.bitable.table.create"
-        | "feishu.bitable.table.patch"
-        | "feishu.bitable.table.batch_create" => &["feishu", "bitable", "table", "write"],
-        "feishu.bitable.record.create"
-        | "feishu.bitable.record.update"
-        | "feishu.bitable.record.delete"
-        | "feishu.bitable.record.batch_create"
-        | "feishu.bitable.record.batch_update"
-        | "feishu.bitable.record.batch_delete" => &["feishu", "bitable", "write"],
-        "feishu.bitable.field.list" => &["feishu", "bitable", "field", "read"],
-        "feishu.bitable.field.create"
-        | "feishu.bitable.field.update"
-        | "feishu.bitable.field.delete" => &["feishu", "bitable", "field", "write"],
-        "feishu.bitable.view.get" | "feishu.bitable.view.list" => {
-            &["feishu", "bitable", "view", "read"]
-        }
-        "feishu.bitable.view.create" | "feishu.bitable.view.patch" => {
-            &["feishu", "bitable", "view", "write"]
-        }
-        "feishu.calendar.freebusy" | "feishu.calendar.list" | "feishu.calendar.primary.get" => {
-            &["feishu", "calendar", "read"]
-        }
-        "feishu.card.update" => &["feishu", "card", "update", "callback"],
-        "feishu.doc.read" => &["feishu", "docs", "read"],
-        "feishu.doc.create" | "feishu.doc.append" => &["feishu", "docs", "write"],
-        "feishu.messages.get" | "feishu.messages.history" | "feishu.messages.search" => {
-            &["feishu", "messages", "read"]
-        }
-        #[cfg(feature = "tool-file")]
-        "feishu.messages.resource.get" => &["feishu", "messages", "resource", "file"],
-        "feishu.messages.send" | "feishu.messages.reply" => &["feishu", "messages", "write"],
-        "feishu.whoami" => &["feishu", "identity", "read"],
-        "tool.search" => &["core", "discover", "search"],
-        "tool.invoke" => &["core", "dispatch", "invoke"],
-        "config.import" => &["config", "import", "migration", "workspace", "legacy"],
-        "external_skills.fetch" => &["skills", "download", "external", "fetch"],
-        "external_skills.resolve" => &["skills", "resolve", "normalize", "external"],
-        "external_skills.search" => &["skills", "search", "inventory", "discover"],
-        "external_skills.recommend" => &["skills", "recommend", "inventory", "discover"],
-        "external_skills.source_search" => &["skills", "search", "discover", "external"],
-        "external_skills.inspect" => &["skills", "inspect", "metadata"],
-        "external_skills.install" => &["skills", "install", "package"],
-        "external_skills.invoke" => &["skills", "invoke", "instructions"],
-        "external_skills.list" => &["skills", "list", "discover"],
-        "external_skills.policy" => &["skills", "policy", "security"],
-        "external_skills.remove" => &["skills", "remove", "uninstall"],
-        "browser.companion.session.start"
-        | "browser.companion.navigate"
-        | "browser.companion.snapshot"
-        | "browser.companion.wait"
-        | "browser.companion.session.stop" => &["browser", "companion", "session", "read"],
-        "browser.companion.click" | "browser.companion.type" => {
-            &["browser", "companion", "write", "approval"]
-        }
-        "http.request" => &["http", "request", "web", "network", "external"],
-        "file.read" => &["file", "read", "filesystem", "repo"],
-        "glob.search" => &[
-            "file",
-            "search",
-            "glob",
-            "filesystem",
-            "repo",
-            "directory",
-            "folder",
-            "list",
-            "browse",
-        ],
-        "content.search" => &["file", "search", "content", "filesystem", "repo"],
-        "memory_search" => &["memory", "search", "recall", "durable", "workspace"],
-        "memory_get" => &["memory", "read", "recall", "durable", "workspace"],
-        "file.write" => &["file", "write", "filesystem"],
-        "file.edit" => &["file", "edit", "filesystem", "exact", "replace"],
-        "shell.exec" => &["shell", "command", "process", "exec"],
-        "bash.exec" => &["bash", "command", "process", "exec"],
-        "provider.switch" => &["provider", "switch", "model", "runtime"],
-        "delegate" | "delegate_async" => &["session", "delegate", "child"],
-        "session_tool_policy_status" | "session_tool_policy_set" | "session_tool_policy_clear" => {
-            &["session", "policy", "tools", "security"]
-        }
-        "session_archive" | "session_cancel" | "session_events" | "session_recover"
-        | "session_status" | "session_wait" | "sessions_history" | "sessions_list" => {
-            &["session", "history", "runtime"]
-        }
-        "session_continue" => &["session", "continue", "delegate", "child"],
-        "sessions_send" => &["session", "message", "channel"],
-        "web.search" => &["web", "search", "discover", "external"],
-        _ => &[],
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -6348,6 +3711,65 @@ mod tests {
 
         assert!(description.contains("channel-backed"));
         assert!(description.contains("Matrix"));
+    }
+
+    #[test]
+    fn session_tool_policy_set_definition_surfaces_runtime_narrowing_shape() {
+        let descriptor = tool_catalog()
+            .descriptor("session_tool_policy_set")
+            .expect("session_tool_policy_set descriptor");
+        let definition = descriptor.provider_definition();
+        let runtime_narrowing =
+            &definition["function"]["parameters"]["properties"]["runtime_narrowing"];
+
+        assert_eq!(runtime_narrowing["type"], "object");
+        assert!(
+            runtime_narrowing["properties"]["browser"]["properties"]["max_sessions"].is_object()
+        );
+        assert!(
+            runtime_narrowing["properties"]["web_fetch"]["properties"]["allowed_domains"]
+                .is_object()
+        );
+    }
+
+    #[test]
+    fn delegate_definitions_surface_shared_and_worktree_isolation_modes() {
+        let catalog = tool_catalog();
+
+        for tool_name in ["delegate", "delegate_async"] {
+            let descriptor = catalog.descriptor(tool_name).expect("delegate descriptor");
+            let definition = descriptor.provider_definition();
+            let isolation =
+                &definition["function"]["parameters"]["properties"]["isolation"]["enum"];
+
+            assert_eq!(*isolation, json!(["shared", "worktree"]));
+        }
+    }
+
+    #[test]
+    fn external_skills_policy_definition_surfaces_update_controls() {
+        let descriptor = tool_catalog()
+            .descriptor("external_skills.policy")
+            .expect("external_skills.policy descriptor");
+        let definition = descriptor.provider_definition();
+        let properties = &definition["function"]["parameters"]["properties"];
+
+        assert_eq!(properties["action"]["enum"], json!(["get", "set", "reset"]));
+        assert!(properties["policy_update_approved"].is_object());
+        assert!(properties["allowed_domains"].is_object());
+        assert!(properties["blocked_domains"].is_object());
+    }
+
+    #[cfg(feature = "tool-browser")]
+    #[test]
+    fn browser_companion_type_definition_requires_session_selector_and_text() {
+        let descriptor = tool_catalog()
+            .descriptor("browser.companion.type")
+            .expect("browser.companion.type descriptor");
+        let definition = descriptor.provider_definition();
+        let required = &definition["function"]["parameters"]["required"];
+
+        assert_eq!(required, &json!(["session_id", "selector", "text"]));
     }
 
     #[cfg(feature = "feishu-integration")]
